frontdesk-be/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
av==14.3.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
email_validator==2.2.0
eval_type_backport==0.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.3.2
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-speech==2.32.0
google-cloud-texttospeech==2.26.0
google-genai==1.3.0
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
livekit==1.0.6
livekit-agents==0.12.20
livekit-api==0.8.2
livekit-plugins-cartesia==0.4.11
livekit-plugins-deepgram==0.7.3
livekit-plugins-google==0.11.3
livekit-plugins-noise-cancellation==0.2.1
livekit-plugins-openai==0.12.3
livekit-plugins-silero==0.7.5
livekit-plugins-turn-detector==0.4.4
livekit-protocol==0.9.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
nest-asyncio==1.6.0
numpy==2.2.5
onnxruntime==1.21.1
openai==1.77.0
packaging==25.0
pillow==11.2.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.4
rsa==4.9.1
safetensors==0.5.3
shellingham==1.5.4
sniffio==1.3.1
sounddevice==0.5.1
SQLAlchemy==2.0.40
starlette==0.46.2
sympy==1.14.0
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.51.3
typer==0.15.3
types-protobuf==4.25.0.20240417
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==14.2
yarl==1.20.0
